Avance_1/requirements.txt
pandas
sqlalchemy
psycopg2-binary
python-dotenv
jupyter
jupyterlab
matplotlib
seaborn
plotly
